+++ modcc/expression.hpp
@@ -11,7 +11,8 @@
 #include "identifier.hpp"
 #include "memop.hpp"
 #include "scope.hpp"
-#include "modccutil.hpp"
+
+#include "io/pprintf.hpp"
 
 class Visitor;
