+++ modcc/functioninliner.cpp
@@ -2,7 +2,6 @@
 
 #include "error.hpp"
 #include "functioninliner.hpp"
-#include "modccutil.hpp"
 #include "errorvisitor.hpp"
 
 expression_ptr inline_function_call(Expression* e)
