+++ modcc/lexer.cpp
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "lexer.hpp"
-#include "modccutil.hpp"
+#include "io/pprintf.hpp"
 
 // helpers for identifying character types
 inline bool in_range(char c, char first, char last) {
