--- modcc/printer/backends/simd.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "backends/avx2.hpp"
-#include "backends/avx512.hpp"
